Revert "wrhg"
This reverts commit 6af93558a71c485bbd21280083f0c0a6bda8824d.

+++ Assets/Script/Zombie/NewAI/EnemyAI.cs
@@ -213,7 +213,7 @@ public class EnemyAI : MonoBehaviour
     {
         startingHealth = newHealth;
     }
-    public void setChasingRange(float chasingRange)
+    public void setChasingRange(int chasingRange)
     {
         Debug.Log("setChasingRange work");
         this.chasingRange = chasingRange;
